+++ src/components/common/applicantAccordions/ApplicantAccordion/ApplicantAccordion.style.ts
