+++ src/types/clients.ts
@@ -1,0 +1,6 @@
+export type applicantdata = {
+  applicantId: number;
+  nickname: string;
+  profileImage: string;
+  applicationDate: string;
+};
